Allow to get handlers programmatically
Console output is put on the cli, making the library to return the raw
handler objects since libraries should not use the console at all.
