+++ index.js
@@ -50,51 +50,17 @@ process.binding = function (name) {
 core.globalTimeouts()
 
 module.exports = function () {
-  var handles = process._getActiveHandles()
   var unknown = 0
   var known = []
 
-  handles.forEach(function (handle) {
+  process._getActiveHandles().forEach(function (handle) {
     var stacks = findStacks(handle, 0)
-    if (stacks) {
-      known.push(stacks)
-      return
-    }
+    if (stacks) return known.push(stacks)
 
     unknown++
   })
 
-  console.error('There are %d known handle(s) keeping the process running and %d unknown', known.length, unknown)
-  console.error('Known handles:\n')
-  known.forEach(function (obj, i) {
-    var stacks = obj.stacks
-
-    stacks = stacks.filter(function (s) {
-      return s.getFileName().indexOf(require('path').sep) > -1
-    })
-
-    console.error('# %s', obj.wrapped.name)
-
-    if (!stacks[0]) {
-      console.error('(unknown stack trace)')
-    } else {
-      var padding = ''
-      stacks.forEach(function (s) {
-        var pad = (s.getFileName() + ':' + s.getLineNumber()).replace(/./g, ' ')
-        if (pad.length > padding.length) padding = pad
-      })
-      stacks.forEach(function (s) {
-        var prefix = s.getFileName() + ':' + s.getLineNumber()
-        try {
-          var src = require('fs').readFileSync(s.getFileName(), 'utf-8').split(/\n|\r\n/)
-          console.error(prefix + padding.slice(prefix.length) + ' - ' + src[s.getLineNumber() - 1].trim())
-        } catch (e) {
-          console.error(prefix + padding.slice(prefix.length))
-        }
-      })
-    }
-    console.error()
-  })
+  return {known: known, unknown: unknown}
 }
 
 function findStacks (obj, depth) {
